+++ telephony-core/src/main/java/telephony/core/service/impl/TaxServiceImpl.java
@@ -1,16 +1,9 @@
 package telephony.core.service.impl;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Date;
-import java.util.List;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
-
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import telephony.core.dao.TaxDao;
 import telephony.core.entity.jpa.Tax;
 import telephony.core.service.SessionService;
@@ -28,9 +21,14 @@ import telephony.core.service.dto.response.TaxEditResponse;
 import telephony.core.service.dto.response.TaxFetchResponse;
 import telephony.core.service.exception.SessionServiceException;
 
-/**
- * asd.
- */
+import java.util.ArrayList;
+import java.util.List;
+
+import telephony.core.service.dto.response.Error;
+
+import static telephony.core.assertion.CommonAssertions.*;
+
+
 public class TaxServiceImpl
 extends AbstractBasicService<Tax>
 implements TaxService {
@@ -52,10 +50,38 @@ implements TaxService {
 		return taxDao.count();
 	}
 
+	// TODO extract to validator
+	private boolean validate(TaxFetchRequest request, List<Error> errors) {
+
+		if (isEmpty(request.getSessionId())) {
+			errors.add(Error.create("sessionId", "sessionId cannot be empty"));
+		}
+
+		if (isEmpty(request.getUsername())) {
+			errors.add(Error.create("username", "username cannot be empty"));
+		}
+
+		if (isNull(request.getFilters())) {
+			errors.add(Error.create("filters", "filters cannot be null"));
+		}
+
+		return errors.size() == 0;
+	}
+
 	@Override
 	@Transactional
 	public TaxFetchResponse fetch(TaxFetchRequest request) throws SessionServiceException {
+
 		logger.info("TaxServiceImpl.fetch starts");
+		TaxFetchResponse resp = new TaxFetchResponse();
+		List<Error> errors = getEmptyErrors();
+
+		if (!validate(request,errors)) {
+			resp.setErrors(errors);
+			resp.setMessage("validationError");
+			resp.setSuccess(false);
+			return resp;
+		}
 
 		if (logger.isDebugEnabled()) {
 			logger.debug("params : [ request : {} ]", request);
@@ -70,47 +96,106 @@ implements TaxService {
 			taxez.add(taxConverter.toDto(entity));
 		}
 
-		TaxFetchResponse resp = new TaxFetchResponse();
-		resp.setMessage(""); // TODO : add localized msg
+		resp.setMessage("operation performed successfully"); // TODO : add localized msg
 		resp.setSuccess(true);
 		resp.setTaxes(taxez);
 
 		return resp;
 	}
 
+	// TODO extract to validator
+	private boolean validate(TaxAddRequest request, List<Error> errors) {
+
+		if (isEmpty(request.getSessionId())) {
+			errors.add(Error.create("sessionId", "sessionId cannot be empty"));
+		}
+
+		if (isEmpty(request.getUsername())) {
+			errors.add(Error.create("username", "username cannot be empty"));
+		}
+
+		if (isNull(request.getTaxDto())) {
+			errors.add(Error.create("taxDto", "taxDto cannot be empty"));
+		}
+
+		// TODO add more cases
+
+		return errors.size() == 0;
+	}
+
 	@Override
 	@Transactional
 	public TaxAddResponse add(TaxAddRequest request) throws SessionServiceException {
+
 		logger.info("TaxServiceImpl.add starts");
+		TaxAddResponse resp = new TaxAddResponse();
+		List<Error> errors = getEmptyErrors();
+
+		if (!validate(request,errors)) {
+			resp.setErrors(errors);
+			resp.setMessage("validationError");
+			resp.setSuccess(false);
+			return resp;
+		}
 
 		if (logger.isDebugEnabled()) {
 			logger.debug("params : [ taxDto : {} ]", request.getTaxDto());
 		}
 
-		sessionService.validate(request.getSessionDto()); // TODO add validation
+		sessionService.validate(request.getSessionDto());
 
 		TaxDto dto = request.getTaxDto();
 		Tax entity = taxConverter.toEntity(dto);
 
 		taxDao.save(entity);
 
-		TaxAddResponse resp = new TaxAddResponse();
 		resp.setSuccess(true);
-		resp.setMessage(""); // TODO : add localized msg
+		resp.setMessage("operation performed successfully"); // TODO : add localized msg
 
 		return resp;
+	}
+
+	// TODO extract to validator
+	private boolean validate(TaxEditRequest request, List<Error> errors) {
+
+		if (isEmpty(request.getSessionId())) {
+			errors.add(Error.create("sessionId", "sessionId cannot be empty"));
+		}
+
+		if (isEmpty(request.getUsername())) {
+			errors.add(Error.create("username", "username cannot be empty"));
+		}
+
+		if (isNull(request.getTaxDto())) {
+			errors.add(Error.create("taxDto", "taxDto cannot be null"));
+			return false;
+		}
+
+		// TODO add more cases
+
+		return errors.size() == 0;
 	}
 
 	@Override
 	@Transactional
 	public TaxEditResponse edit(TaxEditRequest request) throws SessionServiceException {
+
 		logger.info("TaxServiceImpl.edit starts");
+		TaxEditResponse resp = new TaxEditResponse();
+		List<Error> errors = getEmptyErrors();
+
+		if (!validate(request,errors)) {
+			resp.setErrors(errors);
+			resp.setMessage("validationError");
+			resp.setSuccess(false);
+			return resp;
+		}
 
 		if (logger.isDebugEnabled()) {
 			logger.debug("params : [ editTax : {} ]", request.getTaxDto());
 		}
 
-		sessionService.validate(request.getSessionDto()); // TODO add vlaidation
+		sessionService.validate(request.getSessionDto());
 
 		TaxDto dto = request.getTaxDto();
 		Tax tax = taxDao.findById(dto.getId());
@@ -119,29 +204,54 @@ implements TaxService {
 
 		taxDao.saveOrUpdate(tax);
 
-		TaxEditResponse resp = new TaxEditResponse();
-		resp.setMessage(""); // TODO : add localized msg
+		resp.setMessage("operation performed successfully"); // TODO : add localized msg
 		resp.setSuccess(true);
 
 		return resp;
 
 	}
 
+	// TODO extract to validator
+	private boolean validate(TaxDeleteRequest request, List<Error> errors) {
+
+		if (isEmpty(request.getSessionId())) {
+			errors.add(Error.create("sessionId", "sessionId cannot be empty"));
+		}
+
+		if (isEmpty(request.getUsername())) {
+			errors.add(Error.create("username", "username cannot be empty"));
+		}
+
+		if (isEmpty(request.getTaxId())) {
+			errors.add(Error.create("taxId", "taxId cannot by empty"));
+		}
+
+		return errors.size() == 0;
+	}
+
 	@Override
 	@Transactional
 	public TaxDeleteResponse delete(TaxDeleteRequest request) throws SessionServiceException {
+
 		logger.info("TaxServiceImpl.delete starts");
+		TaxDeleteResponse resp = new TaxDeleteResponse();
+		List<Error> errors = getEmptyErrors();
+
+		if (!validate(request,errors)) {
+			resp.setMessage("validationError");
+			resp.setSuccess(false);
+			resp.setErrors(errors);
+			return resp;
+		}
 
 		if (logger.isDebugEnabled()) {
 			logger.debug("params : [ taxId : {} ] ", request.getTaxId());
 		}
 
-		sessionService.validate(request.getSessionDto()); // TODO add validation
-
+		sessionService.validate(request.getSessionDto());
 		taxDao.removeById(request.getTaxId());
 
-		TaxDeleteResponse resp = new TaxDeleteResponse();
-		resp.setMessage(""); // TODO add localized msg
+		resp.setMessage("operation performed successfully"); // TODO add localized msg
 		resp.setSuccess(true);
 		return resp;
 	}
